refactor(mcp): move MCP implementation to extension
