+++ src/mcp/index.ts
@@ -1,48 +1,4 @@
 #!/usr/bin/env node
-/**
- * MCP Server for Claude Desktop
- * Provides memory search tools via Model Context Protocol
- */
-
-import { Server } from '@modelcontextprotocol/sdk/server/index.js';
-import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
-import {
-  ListToolsRequestSchema,
-  CallToolRequestSchema
-} from '@modelcontextprotocol/sdk/types.js';
-import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
-
-import { tools } from './tools.js';
-import { handleToolCall } from './handlers.js';
-
-const server = new Server(
-  {
-    name: 'claude-memory-layer-mcp',
-    version: '1.0.0'
-  },
-  {
-    capabilities: {
-      tools: {}
-    }
-  }
-);
-
-// Tool listing handler
-server.setRequestHandler(ListToolsRequestSchema, async () => {
-  return { tools };
-});
-
-// Tool call handler
-server.setRequestHandler(CallToolRequestSchema, async (request): Promise<CallToolResult> => {
-  const { name, arguments: args } = request.params;
-  return handleToolCall(name, args || {});
-});
-
-// Start server
-async function main() {
-  const transport = new StdioServerTransport();
-  await server.connect(transport);
-  console.error('claude-memory-layer MCP server started');
-}
-
-main().catch(console.error);
+// Compatibility entrypoint. MCP is an optional extension, but existing package
+// and local paths still resolve through src/mcp during the strangler migration.
+import '../extensions/mcp/index.js';
